package/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.10.0
httpx[socks]==0.27.0
asyncio==3.4.3
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
alembic==1.13.1
redis==5.0.1
aioredis==2.0.1
sse-starlette==3.0.3
pytest==8.3.5
pytest-asyncio==0.25.3
psycopg[binary]==3.2.6

# Word 格式化模块依赖
mistune>=3.0.0
python-docx>=1.1.0
lxml>=4.9.0
